refactor(tools): move test helpers to src/test_support.rs
The five `#[cfg(test)]` helpers in `src/tools.rs` (check_func_space,
check_metrics, metrics_verbatim, assert_child_space_kind, child_space)
are used by every per-metric test module, but they lived in a
production file that the self-scan gate walks. Their doc comments and
`#[cfg(test)]` attribute lines are not pruned by `exclude_tests`, so
each new shared helper cost gated `loc.sloc` budget and another
`.bca-baseline.toml` ratchet.

Move all five into a new `#[cfg(test)] mod test_support`, excluded from
the walk via `.bcaignore`. `src/tools.rs` drops from 845 to 749 physical
lines and from 779 to 748 gated `loc.sloc` — under the 760 soft limit,
so its `<file> loc.sloc` baseline entry is retired rather than
ratcheted. `assert_child_space_kind` now delegates to `child_space`
instead of repeating the lookup, and both are `#[track_caller]` so a
failure points at the test rather than the helper. The crate-root
`pub(crate) use crate::tools::check_func_space` re-export had a single
consumer and is gone.

Also correct `bca.toml`'s `exclude_tests` comment: it claimed the
option does not lower `loc.sloc`, which #722 fixed. Measured on
`src/tools.rs` before this change: 779 with the manifest, 845 with
`--no-config`, root span `1..845` either way. The pruned node's rows
are subtracted; its attribute line, doc comment, and surrounding
blanks still count.

Fixes #1066
