+++ src/metrics/loc.rs
@@ -735,7 +735,7 @@ mod typescript;
     clippy::too_many_lines
 )]
 mod tests {
-    use crate::tools::{check_metrics, metrics_verbatim};
+    use crate::test_support::{check_metrics, metrics_verbatim};
 
     use super::*;
