+++ src/metrics/mi.rs
@@ -190,7 +190,7 @@ implement_metric_trait!(
     clippy::too_many_lines
 )]
 mod tests {
-    use crate::tools::check_metrics;
+    use crate::test_support::check_metrics;
 
     use super::*;
