+++ src/metrics/nargs.rs
@@ -551,7 +551,7 @@ impl NArgs for GroovyCode {
     clippy::too_many_lines
 )]
 mod tests {
-    use crate::tools::check_metrics;
+    use crate::test_support::check_metrics;
 
     use super::*;
