+++ src/metrics/nom.rs
@@ -285,7 +285,7 @@ implement_metric_trait!(
     clippy::too_many_lines
 )]
 mod tests {
-    use crate::tools::check_metrics;
+    use crate::test_support::check_metrics;
 
     use super::*;
