+++ src/tools.rs
@@ -745,101 +745,5 @@ pub(crate) fn intense_color(stdout: &mut dyn WriteColor, color: Color) -> std::i
 }
 
 #[cfg(test)]
-pub(crate) fn check_func_space<T: crate::ParserTrait, F: Fn(crate::FuncSpace)>(
-    source: &str,
-    filename: &str,
-    check: F,
-) {
-    let path = std::path::PathBuf::from(filename);
-    // Mirror the CRLF/CR normalisation that read_file_with_eol applies via normalize_line_endings
-    let normalized = source.replace("\r\n", "\n").replace('\r', "\n");
-    let mut trimmed_bytes = normalized.trim_end().trim_matches('\n').as_bytes().to_vec();
-    trimmed_bytes.push(b'\n');
-    let parser = T::new(trimmed_bytes, &path, None);
-    let func_space = crate::spaces::metrics_inner(
-        &parser,
-        path.to_str().map(str::to_owned),
-        crate::MetricsOptions::default(),
-    )
-    .unwrap();
-
-    check(func_space);
-}
-
-#[cfg(test)]
-pub(crate) fn check_metrics<T: crate::ParserTrait>(
-    source: &str,
-    filename: &str,
-    check: fn(crate::CodeMetrics) -> (),
-) {
-    check_func_space::<T, _>(source, filename, |func_space| {
-        check(func_space.metrics.clone());
-    });
-}
-
-/// Analyses `source` **byte-for-byte** and returns its metrics.
-///
-/// Use this, not [`check_metrics`], when a test's input must reach the
-/// parser unaltered: `check_func_space` normalises CRLF and trims then
-/// re-appends a trailing newline, so a construct ending at EOF is
-/// unreachable through it and such a test passes vacuously (#1051). It
-/// also returns a value, which `check_metrics`' bare `fn` callback
-/// cannot. Restrict `options` in timing-sensitive tests so an unrelated
-/// metric's cost cannot dominate and misattribute a regression.
-#[cfg(test)]
-#[track_caller]
-pub(crate) fn metrics_verbatim(
-    lang: crate::LANG,
-    source: &[u8],
-    options: crate::MetricsOptions,
-) -> crate::CodeMetrics {
-    crate::analyze(crate::Source::new(lang, source), options)
-        .expect("verbatim source must analyse")
-        .metrics
-        .clone()
-}
-
-/// Asserts that `func_space` has a direct child space named `name` and that
-/// its `kind` matches `expected`.
-///
-/// Used by annotation-type / class / interface tests that need to verify
-/// the structural FuncSpace tree (not just metric values), since vacuous
-/// metric assertions can pass even when `is_func_space` has been reverted
-/// for the node kind under test.
-#[cfg(test)]
-pub(crate) fn assert_child_space_kind(
-    func_space: &crate::FuncSpace,
-    name: &str,
-    expected: crate::SpaceKind,
-) {
-    let child = func_space
-        .spaces
-        .iter()
-        .find(|s| s.name.as_deref() == Some(name))
-        .unwrap_or_else(|| panic!("expected a child FuncSpace named {name:?}"));
-    assert_eq!(
-        child.kind, expected,
-        "child FuncSpace {name:?} kind: got {:?}, expected {:?}",
-        child.kind, expected,
-    );
-}
-
-/// Returns the direct child [`FuncSpace`][crate::FuncSpace] named `name`,
-/// panicking if absent. Lets per-space-metric tests assert accessors on the
-/// owning class / interface space (where the value is nonzero) rather than on
-/// the always-zero file-unit root.
-#[cfg(test)]
-pub(crate) fn child_space<'a>(
-    func_space: &'a crate::FuncSpace,
-    name: &str,
-) -> &'a crate::FuncSpace {
-    func_space
-        .spaces
-        .iter()
-        .find(|s| s.name.as_deref() == Some(name))
-        .unwrap_or_else(|| panic!("expected a child FuncSpace named {name:?}"))
-}
-
-#[cfg(test)]
 #[path = "tools_tests.rs"]
 mod tests;
